+++ test/flushbar_test.dart
@@ -23,6 +23,17 @@ void main() {
     expect(flushbar.boxShadows, null);
     expect(flushbar.backgroundGradient, null);
     expect(flushbar.mainButton, null);
+    expect(flushbar.borderRadius, 0.0);
+    expect(flushbar.borderWidth, 1.0);
+    expect(flushbar.borderColor, null);
+    expect(flushbar.padding.left, 16);
+    expect(flushbar.padding.right, 16);
+    expect(flushbar.padding.top, 16);
+    expect(flushbar.padding.bottom, 16);
+    expect(flushbar.margin.left, 0);
+    expect(flushbar.margin.right, 0);
+    expect(flushbar.margin.top, 0);
+    expect(flushbar.margin.bottom, 0);
     expect(flushbar.onTap, null);
     expect(flushbar.isDismissible, true);
     expect(flushbar.dismissDirection, FlushbarDismissDirection.VERTICAL);